feat: f32 precision throughout — F32x16 SIMD MatVec, no f64
Energy vector, ResonanceDto, BusDto all f32 now.
u8 distance table has 8 bits. f32 has 24-bit mantissa.
No f64 needed — saves CPU for real perturbation, not artificial precision.

MatVec cycle: 4× F32x16 = 64 elements per iteration.
4096 / 64 = 64 iterations per row. Native FMA.

Next: BF16 energy (7-bit mantissa matches u8 data).
AMX path: u8×u8→i32 systolic for maximum throughput.

76 tests pass.

+++ crates/thinking-engine/examples/text_to_thought.rs
@@ -116,7 +116,7 @@ fn main() {
         let active = engine.active_count(0.001);
 
         // Top-5 peaks
-        let mut peaks: Vec<(usize, f64)> = engine.energy.iter()
+        let mut peaks: Vec<(usize, f32)> = engine.energy.iter()
             .enumerate()
             .filter(|(_, &e)| e > 0.001)
             .map(|(i, &e)| (i, e))
@@ -175,7 +175,7 @@ fn main() {
             let a = dominant_atoms[i].0 as usize;
             let b = dominant_atoms[j].0 as usize;
             let dist = table_data[a * n + b];
-            let cos = (dist as f64 / 255.0) * 2.0 - 1.0;
+            let cos = (dist as f32 / 255.0) * 2.0 - 1.0;
             println!("    S{} ↔ S{}: u8={:>3} cos={:.3}  [{} ↔ {}]",
                 i+1, j+1, dist, cos,
                 &dominant_atoms[i].1,
